[Refactor] Move store path parsing to platform layer
