+++ be/src/platform/CMakeLists.txt
@@ -21,6 +21,7 @@ set(PLATFORM_FILES
   download_util.cpp
   platform_env.cpp
   platform_metrics.cpp
+  store_path.cpp
   thrift_rpc_helper.cpp
   udf_downloader.cpp
 )
