+++ be/src/tools/CMakeLists.txt
@@ -21,6 +21,8 @@ add_library(Tools STATIC
     meta_tool.cpp
 )
 
+target_link_libraries(Tools PRIVATE Platform)
+
 # libmockjvm will be build as shared lib. we use it to mock a libjvm.so to 
 # make sure BE can start without real java runtime
 add_library(mockjvm SHARED mock_jvm.c)
