+++ be/test/platform/CMakeLists.txt
@@ -17,6 +17,7 @@ set(PLATFORM_TEST_FILES
     download_util_test.cpp
     platform_env_test.cpp
     platform_metrics_test.cpp
+    store_path_test.cpp
     thrift_rpc_helper_test.cpp
 )
