+++ be/test/storage/lake/replication_txn_manager_test.cpp
@@ -29,6 +29,7 @@
 #include "fs/fs.h"
 #include "fs/fs_util.h"
 #include "fs/key_cache.h"
+#include "platform/store_path.h"
 #include "runtime/exec_env.h"
 #include "storage/chunk_helper.h"
 #include "storage/delta_column_group.h"
@@ -41,7 +42,6 @@
 #include "storage/lake/transactions.h"
 #include "storage/lake/update_manager.h"
 #include "storage/olap_define.h"
-#include "storage/options.h"
 #include "storage/replication_utils.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
