+++ be/test/storage/lake/tablet_manager_test.cpp
@@ -30,6 +30,7 @@
 #include "common/config_storage_fwd.h"
 #include "fs/fs.h"
 #include "fs/fs_util.h"
+#include "platform/store_path.h"
 #include "storage/lake/fixed_location_provider.h"
 #include "storage/lake/join_path.h"
 #include "storage/lake/location_provider.h"
@@ -37,7 +38,6 @@
 #include "storage/lake/options.h"
 #include "storage/lake/update_manager.h"
 #include "storage/lake/versioned_tablet.h"
-#include "storage/options.h"
 #include "storage/rowset/segment.h"
 #include "storage/tablet_schema.h"
 #include "test_util.h"
